requirements.txt
requests
six
pyYAML
datasets_lib>=1.2.2
confobj>=1.2.0